+++ app/lib/shared/versions.dart
@@ -22,7 +22,7 @@ final RegExp runtimeVersionPattern = RegExp(r'^\d{4}\.\d{2}\.\d{2}$');
 /// when the version switch happens.
 const acceptedRuntimeVersions = <String>[
   // The current [runtimeVersion].
-  '2023.03.29',
+  '2023.04.11',
   // Fallback runtime versions.
   '2023.03.13',
   '2023.03.07',
@@ -54,10 +54,10 @@ bool shouldGCVersion(String version) =>
 
 // keep in-sync with SDK version in .mono_repo.yml and Dockerfile
 final String runtimeSdkVersion = '2.19.0';
-final String toolStableDartSdkVersion = '2.19.4';
-final String toolStableFlutterSdkVersion = '3.7.7';
-final String toolPreviewDartSdkVersion = '3.0.0-305.0.dev';
-final String toolPreviewFlutterSdkVersion = '3.8.0-10.1.pre';
+final String toolStableDartSdkVersion = '2.19.6';
+final String toolStableFlutterSdkVersion = '3.7.10';
+final String toolPreviewDartSdkVersion = '3.0.0-400.0.dev';
+final String toolPreviewFlutterSdkVersion = '3.9.0-0.2.pre';
 
 final semanticToolStableDartSdkVersion =
     Version.parse(toolStableDartSdkVersion);