+++ CHANGELOG.md
@@ -2,7 +2,11 @@ Important changes to data models, configuration, and migrations between each
 AppEngine version, listed here to ease deployment and troubleshooting.
 
 ## Next Release (replace with git tag when deployed)
- * Bumped runtimeVersion to `2023.03.29`.
+ * Bumped runtimeVersion to `2023.04.11`.
+ * Upgraded stable Dart analysis SDK to `2.19.6`.
+ * Upgraded stable Flutter analysis SDK to `3.7.10`.
+ * Upgraded preview Dart analysis SDK to `3.0.0-400.0.dev`.
+ * Upgraded preview Flutter analysis SDK to `3.9.0-0.2.pre`.
  * Upgraded dartdoc to `6.2.1`.
  * Upgraded `package:markdown` to `7.0.2`
  * Upgraded material web components to `14.0.0`.